1inch-swap/requirements.txt
web3==5.29.1
python-dotenv==0.20.0
requests==2.27.1
